requirements.txt
PyGithub==2.1.1
scikit-learn==1.3.2
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
halo==0.0.31
termcolor==2.4.0
